+++ src/completion/ValueCompletionItem.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
-import { ProjectCollector } from '../Projects/ProjectCollector';
+import { Solution } from '../Classes/Solution';
+import { Keyword } from '../Constants/Keyword';
 
 export class ValueCompletionItemProvider implements vscode.CompletionItemProvider
 {
@@ -11,6 +12,8 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
     {
         return new Promise<vscode.CompletionItem[]|vscode.CompletionList>((resolve, rejects) =>
         {
+            // only add theses completions values to the list,
+            // when a user has typed a trigger character and not press ALT+ENTER or CTRL+.
             if(context.triggerCharacter === undefined)
             {
                 rejects();
@@ -19,12 +22,13 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
             const list = new Array<vscode.CompletionItem>();
             const valueList = new Array<[string, vscode.CompletionItemKind]>();
 
-            let line = document.lineAt(position).text.trim().toLowerCase();
-
             let inGlobalSection = false;
             let inProjectDependencies = false;
             let InNestedProjects = false;
-            let inConfigurationPlatforms = false;
+            let inProjectConfigurationPlatforms = false;
+            let inSolutionConfigurationPlatforms = false;
+
+            let line = document.lineAt(position).text.trim().toLowerCase();
 
             if(line.startsWith("globalsection"))
             {
@@ -38,7 +42,6 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
             }
             else
             {
-
                 for(let lineNumber = (position.line - 1); lineNumber > 0; lineNumber--)
                 {
                     line = document.lineAt(lineNumber).text.trim().toLowerCase();
@@ -62,10 +65,15 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
                         break;
                     }
 
-                    if(line.startsWith("globalsection(projectconfigurationplatforms)")
-                    || line.startsWith("globalsection(solutionconfigurationplatforms)"))
+                    if(line.startsWith("globalsection(projectconfigurationplatforms)"))
                     {
-                        inConfigurationPlatforms = true;
+                        inProjectConfigurationPlatforms = true;
+                        break;
+                    }
+
+                    if(line.startsWith("globalsection(solutionconfigurationplatforms)"))
+                    {
+                        inSolutionConfigurationPlatforms = true;
                         break;
                     }
 
@@ -77,7 +85,7 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
                 }
             }
 
-            if(inConfigurationPlatforms)
+            if(inSolutionConfigurationPlatforms)
             {
                 valueList.push(['Debug|Any CPU', vscode.CompletionItemKind.Value]);
                 valueList.push(['Debug|ARM', vscode.CompletionItemKind.Value]);
@@ -90,36 +98,98 @@ export class ValueCompletionItemProvider implements vscode.CompletionItemProvide
                 valueList.push(['Release|x64', vscode.CompletionItemKind.Value]);
                 valueList.push(['Release|x86', vscode.CompletionItemKind.Value]);
             }
-            
+
             if(inGlobalSection)
             {
                 valueList.push(['TRUE', vscode.CompletionItemKind.Constant]);
                 valueList.push(['FALSE', vscode.CompletionItemKind.Constant]);
             }
 
-            for(const [name, kind] of valueList)
+            if(inProjectDependencies || InNestedProjects || inProjectConfigurationPlatforms)
             {
-                const completionItem = new vscode.CompletionItem(name, kind);
-                completionItem.insertText = ` ${name}`;
-                list.push(completionItem);
-            }
+                const solution = new Solution(document);
 
-            if(inProjectDependencies || InNestedProjects)
-            {
-                const guidList = new ProjectCollector(document, false).ProjectList;
-
-                for(const project of guidList)
+                if(inProjectDependencies || InNestedProjects)
                 {
-                    const completionItem = new vscode.CompletionItem(project.Name, vscode.CompletionItemKind.Reference);
-                    completionItem.detail = `Guid: ${project.Guid}`
-                    completionItem.insertText = ` {${project.Guid}}`;
-                    completionItem.documentation = `Project type: ${project.GetProjectTypeName()}\n\nPath:\n${project.GetProjectPath()}`;
+                    this.AddProjectsToCompletionList(solution, list);
+                }
 
-                    list.push(completionItem);
+                if(inProjectConfigurationPlatforms)
+                {
+                    this.AddConfigurationsToCompletionList(solution, list);
                 }
             }
 
+            this.AddValuesToCompletionList(valueList, list);
+
             resolve(list);
         });
+    }
+
+    /**
+     * Add all projects the a defined in the solution to the code completions list
+     * @param solution The solution that contains the projects
+     * @param completionList The list for the code completions
+     */
+    private AddProjectsToCompletionList(
+        solution: Solution,
+        completionList: Array<vscode.CompletionItem>): void
+    {
+        for(const project of solution.Projects)
+        {
+            const completionItem = new vscode.CompletionItem(
+                project.Name,
+                vscode.CompletionItemKind.Reference);
+
+            completionItem.detail = `Guid: ${project.Guid}`;
+            completionItem.insertText = ` {${project.Guid}}`;
+
+            completionItem.documentation
+                = `Project type: ${project.GetProjectTypeName()}\n\nPath:\n${project.GetProjectPath()}`;
+
+            completionList.push(completionItem);
+        }
+    }
+
+    /**
+     * Add all configurations that are defined in the solution to the code completions list
+     * @param solution The solution that contains the configurations
+     * @param completionList The list for the code completions
+     */
+    private AddConfigurationsToCompletionList(
+        solution: Solution,
+        completionList: Array<vscode.CompletionItem>): void
+    {
+        const configurationPlatforms = solution.Global?.GlobalSections
+            .find(section => section.Type == Keyword.SolutionConfigurationPlatforms);
+
+        if(configurationPlatforms === undefined)
+        {
+            return;
+        }
+
+        for(const [, , value] of configurationPlatforms.KeyValueList)
+        {
+            const completionItem = new vscode.CompletionItem(value, vscode.CompletionItemKind.Value);
+            completionItem.insertText = ` ${value}`;
+            completionList.push(completionItem);
+        }
+    }
+
+    /**
+     * 
+     * @param valueList The list with all values for the code completion list
+     * @param completionList The list for the code completions
+     */
+    private AddValuesToCompletionList(
+        valueList: Array<[string, vscode.CompletionItemKind]>,
+        completionList: Array<vscode.CompletionItem>): void
+    {
+        for(const [name, kind] of valueList)
+        {
+            const completionItem = new vscode.CompletionItem(name, kind);
+            completionItem.insertText = ` ${name}`;
+            completionList.push(completionItem);
+        }
     }
 }